Various cleanups prompted by Benjy's review.
